{{cookiecutter.project_slug}}/requirements.txt
click==6.6
clickclick==1.2.1
connexion==1.0.129
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
MarkupSafe==0.23
python-dateutil==2.6.0
PyYAML==3.12
requests==2.12.3
six==1.10.0
strict-rfc3339==0.7
swagger-spec-validator==2.0.2
uWSGI==2.0.14
Werkzeug==0.11.11
